core: add atom + world: add createDisc
core
- atom is reactive state primitive
- Tokens are refactored to use atom internally
- Tokens also accept atom as argument
- add effect to subscribe to multiple atoms

world
- add createDisc, this is the first generated Shape in tagl and makes use of atoms

+++ dev/examples/03_world.ts
@@ -23,7 +23,7 @@ plane.color.subscribe((value) => {
   console.log('value', value)
 })
 
-scene.camera.set((camera) => mat4.translate(camera, camera, [0, 0, -1]) as Float32Array)
+scene.camera.set((camera) => mat4.translate(camera, camera, [0, 0, -1]))
 scene.add(plane)
 
 setTimeout(() => {
